feat: create script to generate new proto files
This commit introduces a new script, `create_new_proto_file.sh`,
that automates the creation of new protocol buffer files and
directories. The script takes a proto name as an argument, creates
the corresponding directory and `.proto` file, and updates the
`buf.work.yaml` file to include the new directory. This streamlines
the process of adding new proto definitions to the project.